llvm_tools: fix autosubmit label invocation
I checked the logs for this job, and it noted that the worktree
being created wasn't within a CrOS source tree. Since this script
is meant to be invoked from within a CrOS source tree, swapping the
'dir' arg to the toolchain-utils root should fix this.

BUG=None
TEST=unittests

Change-Id: Ie5ecf970ec1354c75ae8eeb3b37d273160bc2868
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6093017
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>