llvm_tools: add --force-checkout to ready_llvm_branch.py
Iterating with this tool is a bit awkward; when I rebase a patch and
try to run it again, it fails because the LLVM branch I'm trying to
(re)create already exists.

Add `--force-checkout` to wipe out all existing state and start over, so
iteration is easier.

BUG=b:404909262
TEST=Ran the tool with `--force-checkout`; no more warnings about
TEST=losing state or anything :)

Change-Id: Ic5bb4502299f6e00b2c45884c3253591729a0e1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6416644
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>